Move the Calendar from the Treasury tab to the Header
